UPDATE: magnetic sensor synchronization with ADC sampling time,SWAP BLDC channel

+++ Drivers/LIB/Inc/DLPF_lib.h
@@ -11,10 +11,10 @@
 
 typedef struct
 {
-  double result, alpha;
+  float result, alpha;
 }DLPF_HandleTypeDef;
 
-void dlpf_set_alpha (DLPF_HandleTypeDef *hdlpf, double alpha_val);
-double dlpf_get_result (DLPF_HandleTypeDef *hdlpf, double raw_data);
+void dlpf_set_alpha (DLPF_HandleTypeDef *hdlpf, float alpha_val);
+float dlpf_get_result (DLPF_HandleTypeDef *hdlpf, float raw_data);
 
 #endif /* LIB_INC_DLPF_LIB_H_ */
