+++ Drivers/LIB/Scr/DLPF_lib.c
@@ -9,14 +9,14 @@
 
 //============================================================================================
 
-void dlpf_set_alpha (DLPF_HandleTypeDef *hdlpf, double alpha_val)
+void dlpf_set_alpha (DLPF_HandleTypeDef *hdlpf, float alpha_val)
 {
 	hdlpf->alpha = alpha_val;
 }
 
 //============================================================================================
 
-double dlpf_get_result (DLPF_HandleTypeDef *hdlpf, double raw_data)
+float dlpf_get_result (DLPF_HandleTypeDef *hdlpf, float raw_data)
 {
 	hdlpf->result = hdlpf->result * (1.0 - hdlpf->alpha) + raw_data * hdlpf->alpha;
 	return hdlpf->result;
